+++ lib/src/yako_theme_switch.dart
@@ -8,8 +8,8 @@ class YakoThemeSwitch extends StatefulWidget {
   /// You can set the custom width of the switch
   final double width;
 
-  /// Just a callback function to get the changed state of the switch
-  final Function({bool changed}) onChanged;
+  /// Callback invoked with the new state whenever the switch is toggled
+  final ValueChanged<bool> onChanged;
 
   /// The background color of the switch when it is enabled
   final Color? enabledBackgroundColor;
@@ -26,10 +26,11 @@ class YakoThemeSwitch extends StatefulWidget {
   /// The duration of the animation
   final Duration animationDuration;
 
-  /// The border radius of the toggle. Try setting it to 4 and chech how circle now is square
+  /// The border radius of the toggle. Try setting it to 4 to make the circle square
   final double? enabledToggleBorderRadius;
 
   const YakoThemeSwitch({
+    super.key,
     this.enabled = false,
     this.width = 45,
     this.enabledBackgroundColor,
@@ -42,23 +43,23 @@ class YakoThemeSwitch extends StatefulWidget {
   });
 
   @override
-  _YakoThemeSwitchState createState() => _YakoThemeSwitchState();
+  State<YakoThemeSwitch> createState() => _YakoThemeSwitchState();
 }
 
 class _YakoThemeSwitchState extends State<YakoThemeSwitch>
     with SingleTickerProviderStateMixin {
-  late AnimationController animationController;
-  late Animation<double> animation;
-  late bool turnState;
-  late Color enabledBackgroundColor;
-  late Color disabledBackgroundColor;
-  late Color enabledToggleColor;
-  late Color disabledToggleColor;
-  double animationControllerValue = 0.0;
+  late AnimationController _animationController;
+  late Animation<double> _animation;
+  late bool _turnState;
+  late Color _enabledBackgroundColor;
+  late Color _disabledBackgroundColor;
+  late Color _enabledToggleColor;
+  late Color _disabledToggleColor;
+  double _animationValue = 0.0;
 
   @override
   void dispose() {
-    animationController.dispose();
+    _animationController.dispose();
     super.dispose();
   }
 
@@ -66,37 +67,33 @@ class _YakoThemeSwitchState extends State<YakoThemeSwitch>
   void initState() {
     super.initState();
 
-    enabledBackgroundColor =
+    _enabledBackgroundColor =
         widget.enabledBackgroundColor ?? Colors.grey.shade300;
-    disabledBackgroundColor =
-        widget.disabledBackgroundColor ?? Color(0xFF2E386E);
-    enabledToggleColor =
+    _disabledBackgroundColor =
+        widget.disabledBackgroundColor ?? const Color(0xFF2E386E);
+    _enabledToggleColor =
         widget.enabledToggleColor ?? Colors.amberAccent.shade700;
-    disabledToggleColor = widget.disabledToggleColor ?? Color(0xFF70E2FB);
+    _disabledToggleColor =
+        widget.disabledToggleColor ?? const Color(0xFF70E2FB);
 
-    animationControllerValue = widget.enabled ? 1.0 : 0.0;
-    animationController = AnimationController(
-      value: animationControllerValue,
+    _turnState = widget.enabled;
+    _animationValue = widget.enabled ? 1.0 : 0.0;
+
+    _animationController = AnimationController(
+      value: _animationValue,
       vsync: this,
       lowerBound: 0.0,
       upperBound: 1.0,
       duration: widget.animationDuration,
     );
-    animation =
-        CurvedAnimation(parent: animationController, curve: Curves.easeInOut);
+    _animation = CurvedAnimation(
+      parent: _animationController,
+      curve: Curves.easeInOut,
+    );
 
-    animationController.addListener(() {
+    _animationController.addListener(() {
       setState(() {
-        animationControllerValue = animation.value;
-      });
-    });
-    turnState = widget.enabled;
-
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      setState(() {
-        if (turnState) {
-          animationController.forward();
-        }
+        _animationValue = _animation.value;
       });
     });
   }
@@ -105,13 +102,12 @@ class _YakoThemeSwitchState extends State<YakoThemeSwitch>
   void didUpdateWidget(YakoThemeSwitch oldWidget) {
     super.didUpdateWidget(oldWidget);
 
-    // if enabled value is updated, reflect the change in switch
     if (oldWidget.enabled != widget.enabled) {
-      turnState = widget.enabled;
-      if (turnState) {
-        animationController.forward();
+      _turnState = widget.enabled;
+      if (_turnState) {
+        _animationController.forward();
       } else {
-        animationController.reverse();
+        _animationController.reverse();
       }
     }
   }
@@ -119,15 +115,13 @@ class _YakoThemeSwitchState extends State<YakoThemeSwitch>
   @override
   Widget build(BuildContext context) {
     final Color? transitionColor = Color.lerp(
-      disabledBackgroundColor,
-      enabledBackgroundColor,
-      animationControllerValue,
+      _disabledBackgroundColor,
+      _enabledBackgroundColor,
+      _animationValue,
     );
 
     return GestureDetector(
-      onTap: () {
-        toggle(changeState: true);
-      },
+      onTap: _toggle,
       child: Container(
         padding: const EdgeInsets.all(3),
         width: widget.width,
@@ -138,23 +132,21 @@ class _YakoThemeSwitchState extends State<YakoThemeSwitch>
         child: Stack(
           children: <Widget>[
             Transform.translate(
-              offset: Offset((widget.width - 25) * animationControllerValue, 0),
+              offset: Offset((widget.width - 25) * _animationValue, 0),
               child: Transform.rotate(
-                angle: animationControllerValue,
-                child: Container(
+                angle: _animationValue,
+                child: SizedBox(
                   height: 18,
                   width: 18,
-                  alignment: Alignment.center,
                   child: Stack(
                     children: <Widget>[
                       Center(
                         child: Opacity(
-                          opacity:
-                              (1 - animationControllerValue).clamp(0.0, 1.0),
+                          opacity: (1 - _animationValue).clamp(0.0, 1.0),
                           child: SvgPicture.asset(
                             'assets/dark_mode_switch_icon.svg',
                             colorFilter: ColorFilter.mode(
-                              disabledToggleColor,
+                              _disabledToggleColor,
                               BlendMode.srcIn,
                             ),
                             height: 18,
@@ -164,12 +156,13 @@ class _YakoThemeSwitchState extends State<YakoThemeSwitch>
                       ),
                       Center(
                         child: Opacity(
-                          opacity: animationControllerValue.clamp(0.0, 1.0),
+                          opacity: _animationValue.clamp(0.0, 1.0),
                           child: Container(
                             decoration: BoxDecoration(
                               borderRadius: BorderRadius.circular(
-                                  widget.enabledToggleBorderRadius ?? 20),
-                              color: enabledToggleColor,
+                                widget.enabledToggleBorderRadius ?? 20,
+                              ),
+                              color: _enabledToggleColor,
                             ),
                           ),
                         ),
@@ -178,19 +171,22 @@ class _YakoThemeSwitchState extends State<YakoThemeSwitch>
                   ),
                 ),
               ),
-            )
+            ),
           ],
         ),
       ),
     );
   }
 
-  void toggle({bool changeState = false}) {
+  void _toggle() {
     setState(() {
-      if (changeState) turnState = !turnState;
-      turnState ? animationController.forward() : animationController.reverse();
-
-      widget.onChanged(changed: turnState);
+      _turnState = !_turnState;
+      if (_turnState) {
+        _animationController.forward();
+      } else {
+        _animationController.reverse();
+      }
     });
+    widget.onChanged(_turnState);
   }
 }